Update traveler and extract overlap count
We want the overlap information for the RNAcentral pipeline. Currently,
the only way to get that is to parse the log that traveler produces in
verbose mode. To do that, this saves the log file alongside the other
output files and then parses it and creates a .overlaps file with a
single line that is the number of overlaps.