Remove more hardcoded paths
These paths cannot be relied upon as we may be running in the cluster
where we do not have write permissions. To deal with this I have altered
the code to pass the data paths around more cleanly.

Also, I have added several commands to serve as clear entry points. This
is a bit nice (to me) as we can tell what options go where.

The setup logic is moved to the script as well. This may be a bad idea
for Rfam based families, but it is present for now.