Stub command to validate Rfam families
This command will be used to dteect if an Rfam family can be safely
drawn. Some families cannot because their emplates have too many
overlaps to start with. This does not actually implement the logic to do
so (as I don't know what it is yet), but it does create the command so
the pipeline can start to use it. Currently the command will always
write out the given id meaning that it is valid.