Be more aggressive about errors in running programs
Basically we should always have a working cmalign/esl-*/etc. If they do
not work then something is wrong and we should stop. On the other hand
traveler can fail sometimes and that is ok. This also helps me debug
things by failing early, instead of later when things are in a broken
state.