Better deal with certain errors
From running this on all of RNAcentral it seems that it is very common
for some sequences to fail the alimanip stage (reason unclear right
now). Because of this many inputs end up crashing because a single
sequence out of hundreds fails. By treating this error like we do
missing models, log and continue, we should be able to get many more
diagrams.