Tolerate missing traveler files
It seems that in some cases the traveler files are not available, for
example RF_site5 cannot be found in the image. As a result r2dt tries to
build what is needed, but this causes the pipeline to fail when run in a
container because the container file system is read only. This catches
such an error, logs and continues on so other sequences can still be
drawn.