Initial script to convert SVG to JSON for PDBe
We want to convert results of R2DT into a structure that is useful for
PDBe. This is a simple python script to do so. It is pretty simple
currently but more or less works. I have some concerns that this will
not deal well with sequences that have missing 5' or 3' ends as the
resnum property is just sequential numbering and if those parts do not
match up to the modeled ones they will not be drawn. This will break the
number by either making it shifted off by some amount or make it
incomplete.